services/analytics/pyproject.toml
[tool.poetry]
name = "analytics"
version = "0.1.0"
description = "Analytics service for the TODO project"
authors = ["ilia <iliaforwork1993@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11"
fastapi = "^0.111.1"
uvicorn = {extras = ["standard"], version = "^0.30.3"}
motor = "^3.5.0"
python-dotenv = "^1.0.1"
celery = "^5.4.0"
redis = "^5.0.7"
requests = "^2.32.3"
kafka-python = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
